+++ webknossos/examples/learned_segmenter.py
@@ -52,10 +52,8 @@ def main() -> None:
     )  # wk data has dimensions (Channels, X, Y, Z)
     # move channels to last dimension, remove z dimension to match skimage's shape
     X_train = np.moveaxis(np.squeeze(img_data_train), 0, -1)
-    print(volume_annotation.mags[mag].bounding_box)
     Y_train = np.squeeze(volume_annotation.mags[mag].read())
 
-    print(training_data_bbox, X_train.shape, Y_train.shape)
     segmenter.fit(X_train, Y_train)
 
     # Step 4: Use our trained model and predict a class for each pixel in the dataset