+++ webknossos/Changelog.md
@@ -12,10 +12,12 @@ For upgrade instructions, please check the respective *Breaking Changes* section
 ### Breaking Changes
 
 ### Added
+- Added `remove_layer` to the `Dataset` class.
 
 ### Changed
 
 ### Fixed
+- `dataset.upload(layers_to_link=…)`: Fixed a bug where the upload did not complete if layers_to_link contained layers present in uploading dataset.
 
 
 ## [0.9.2](https://github.com/scalableminds/webknossos-libs/releases/tag/v0.9.2) - 2022-02-03